bot/requirements.txt
discord.py>=2.3.2
httpx>=0.27.0
python-dotenv>=1.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
numpy<2

# langchain dependencies
langchain
langchain-core>=0.1.0
langgraph
langchain_huggingface
langchain-deepseek
langchain-mistralai
langchain-openai

duckduckgo-search